Fix CPU topology
Current setting adds the requested CPUs as sockets. While this works, it
is a tad bit in-accurate in the sense there are never machines with more
than 4 sockets. Some applications that depend on CPU topology become
intolerant to this setup. Changing the default code to tag all the
requested CPUs as cores instead of sockets which seems to allow those
applications to continue with their operations. Hopefully the
performance impact is minimal with this change

Signed-off-by: Ganesh Maharaj Mahalingam <ganesh.mahalingam@intel.com>