Increase default shm size
In k8s, the way to increase the size of /dev/shm is to set it as an
EmptyDir and back it with memory from the host. While this works fine
with runc, in kata, we still only get the default of 4K. This is a pure
hack for the current use-case but needs to be fixed.

Signed-off-by: Ganesh Maharaj Mahalingam <ganesh.mahalingam@intel.com>